+++ gulpfile.babel.js
@@ -21,5 +21,6 @@ require('./gulp/cleanTask.js')(gulp, $, gulpConfig);
 require('./gulp/buildTask.js')(gulp, $, gulpConfig);
 require('./gulp/docTask.js')(gulp, $, gulpConfig);
 require('./gulp/wsgTask.js')(gulp, $, gulpConfig);
+require('./gulp/jsTask.js')(gulp, $, gulpConfig);
 
 gulp.task('default', $.taskListing);